Fix #71: crash from REPL after hitting return on last line
Avoid sending more than the requested amount of data to R when handling ReadConsole.

Check input length in R REPL provider and prevent user from executing input that is too long.

(This covers the VS side of things and should be sufficient to fix the issue entirely on all single-byte locales. Separate checks in RHost need to be added as a last line of defense and to handle multibyte case.)