Fix #284: Assertion when resetting REPL Fix #255: Exception in Variable Explorer by killing RHost while VE updates
Add support for cancelable evals, and make them non-cancelable by default.

Properly handle new eval response format for canceled evals.

Do not leak expected WebSocketException from RHost when the host is shutting down.

Automatically re-initialize DebugSession when RSession is disconnected and reconnected, and ensure that it is always properly initialized for all methods that require it.