Fix #114: REPL doesn't work for multiple instances of RTVS
Utilize the new connect mode of the R host process to reverse the connection flow to be from host to VS, and select a random available port to listen on to avoid conflicts.

Abstract away low-level message handling into a new message transport abstraction, to simplify code and permit non-websocket transports (e.g. for testing).