Fix #689: Debugger fails if breakpoints are set on function definition and first line of its body
When injecting more than one breakpoint, use the expression produced by previous injection as input for the next one, instead of using the original expression all the time.

Fix #719: Source info is missing at breakpoint that is on a subexpression of another expression with a breakpoint

Recurse into subexpressions of injected breakpoints when copying srcrefs.