Fix the place where setting overflow flag for grid scroll
Overflow flag tells that computed viewport is smaller than render size. So, it has to walk backward. This happens when window size is getting larger beyond the last element in row or column.
The flag should be set regardless to the number of items added in reversed order - render window could be larget the whole grid.