Fix #767: Stop debugging in infinite loop hangs VS
Add a timeout when waiting for prompt to issue "Q" command, and abort the current interaction as a fallback.

When VS requests to delete breakpoints during engine shutdown, do not wait for the corresponding evals to complete.

Minor refactoring - add missing default values for CancellationToken arguments, clean up usings & reformat.