Fix #1205: Stop attaching logs to send a frown email
Always pop up Explorer with the log file selected, even when using Outlook.

Stop automatically attaching logs to email even when using Outlook, and change the text to give directions on how to do so manually.

Add basic diagnostic info (VS, RTVS and R versions) directly into the precomposed email body, so that it is present even if logs are not attached.

Use a custom non-modal window instead of the stock message box as a mailto: fallback, so that the entire body of the message is included there, and can be easily copy/pasted, in case mailto: handler didn't work for any reason.