Rename "fields" parameters to "properties" everywhere that wasn't renamed before, to match the enum name.
Rename EvaluateAndDescribe to TryEvaluateAndDescribe to signify that the method doesn't throw on error, and clarify the possible type for the returned value.

Reuse EvaluateAndDescribe for a new method that wraps TryEvaluateAndDescribe, and throws exceptions for errors, to simplify error checking code in places where error result is not expected or does not require special treatment beyond bailing out.

Add more comments.