Fix #1725: Support non-system encoding for Source R Script command
Add support for encoding to SourceFile helper methods, and fix encoding handling in rtvs:::debug_source.

Use ITextDocument encoding when available in SourceRScript.

Enable SourceRScript command whenever there's an active REPL window, regardless of whether it is visible or not; and make it visible when command is invoked.

Add tests for SourceRScript.